pyproject.toml
[build-system]
requires=["hatchling"]
build-backend="hatchling.build"

[project]
name="shellforgeai"
version="0.1.0"
requires-python=">=3.12"
dependencies=["typer>=0.12","rich>=13","pydantic>=2","pydantic-settings>=2","pyyaml>=6","httpx>=0.27"]
[project.optional-dependencies]
dev=["pytest>=8","ruff>=0.5","mypy>=1.10","types-PyYAML"]
[project.scripts]
shellforgeai="shellforgeai.cli:app"
sfai="shellforgeai.cli:app"
[tool.ruff]
line-length=100
target-version="py312"
[tool.ruff.lint]
select=["E","F","I","UP","B","SIM"]
[tool.mypy]
strict=true
packages=["shellforgeai"]
[[tool.mypy.overrides]]
module=["tests.*"]
disallow_untyped_defs=false
[tool.pytest.ini_options]
testpaths=["tests"]
addopts="-q"
